Rename validation job
